+++ src/plugins/legacy/vtkDataMesh/interactors/vtkDataMesh4DInteractor.cpp
@@ -50,7 +50,7 @@ vtkDataMesh4DInteractor::vtkDataMesh4DInteractor(medAbstractView* parent): vtkDa
     d->view = dynamic_cast<medAbstractImageView *>(parent);
     d->data = nullptr;
 
-    medVtkViewBackend* backend = static_cast<medVtkViewBackend*>(parent->backend());
+    medVtkViewBackend *backend = static_cast<medVtkViewBackend*>(parent->backend());
     d->view2d = backend->view2D;
     d->view3d = backend->view3D;
 
@@ -132,8 +132,8 @@ void vtkDataMesh4DInteractor::setCurrentTime (double time)
     {
         // Refresh view size in case of resize
         QSize size = d->view->viewWidget()->size();
-        int newSizeX = (int)((double)size.width()/90.0);
-        int newSizeY = (int)((double)size.height()/1.5);
+        int newSizeX = static_cast<int>(size.width()/90.0);
+        int newSizeY = static_cast<int>(size.height()/1.5);
 
         // Display Time + Shift
         if (d->textActor == nullptr)
